Add files via upload
This program calculates the entropy, enthalpy, and Gibb's free energy of 
adsorbates, and transition states within zeolites. It is also capable
of calculating the ideal gas-phase entropy, enthalpy, and Gibb's free energy
of molecules and transition states. The necessary input files are VASP
'OUTCAR' and 'CONTCAR'.

The user is given the choice of two approximations for calculating the entropy
of adsorbed molecules/transition states. 

1) The harmonic oscillator-[https://aiche.onlinelibrary.wiley.com/doi/10.1002/aic.16838]
2) 2D free translator-[https://doi.org/10.1021/jp106536m, https://doi.org/10.1016/j.micromeso.2019.04.058],